Was putting these tests in the wrong file, moved to the right place
