+++ specutils/tests/test_manipulation.py
@@ -9,8 +9,27 @@ from astropy.tests.helper import quantity_allclose
 from ..utils.wcs_utils import gwcs_from_array
 
 from ..spectra import Spectrum1D, SpectralRegion, SpectrumCollection
-from ..manipulation import snr_threshold
+from ..manipulation import snr_threshold, excise_regions, linear_exciser
 
+ def test_true_exciser():
+    np.random.seed(84)
+    spectral_axis = np.linspace(5000,5100,num=100)*u.AA
+    flux = (np.random.randn(100) + 3) * u.Jy
+    spec = Spectrum1D(flux=flux, spectral_axis = spectral_axis)
+    region = SpectralRegion([(5005,5010), (5060,5065)]*u.AA)
+    excised_spec = excise_regions(spec, region)
+
+    assert len(excised_spec.spectral_axis) == len(spec.spectral_axis)-10
+    assert len(excised_spec.flux) == len(spec.flux)-10
+    assert np.isclose(excised_spec.flux.sum(), 274.54139*u.Jy, atol=0.001)
+
+def test_linear_exciser():
+    np.random.seed(84)
+    spectral_axis = np.linspace(5000,5100,num=100)*u.AA
+    flux = (np.random.rand(100)*100) * u.Jy
+    spec = Spectrum1D(flux=flux, spectral_axis = spectral_axis)
+    region = SpectralRegion([(5005,5010), (5060,5065)]*u.AA)
+    excised_spec = excise_regions(spec, region, exciser = linear_exciser)
 
 def test_snr_threshold():
